refactor : 타입 조회 메서드 이름을 of -> from으로 변경, 테스트 케이스 로직 수정, 사용하지 않는 어노테이션 제거
