+++ backend/src/main/java/com/fit/nufit/meal/application/MealDetailService.java
@@ -2,7 +2,7 @@ package com.fit.nufit.meal.application;
 
 import com.fit.nufit.food.domain.Food;
 import com.fit.nufit.food.domain.FoodRepository;
-import com.fit.nufit.food.dto.FoodSimpleResponse;
+import com.fit.nufit.food.dto.response.FoodSimpleResponse;
 import com.fit.nufit.meal.domain.Meal;
 import com.fit.nufit.meal.domain.MealDetail;
 import com.fit.nufit.meal.domain.MealDetailRepository;
