+++ backend/src/main/java/com/fit/nufit/meal/domain/MealType.java
@@ -5,7 +5,7 @@ import com.fit.nufit.meal.exception.NoSuchMealException;
 public enum MealType {
     BREAKFAST, LUNCH, DINNER, SNACK;
 
-    public static MealType of(String value) {
+    public static MealType from(String value) {
         try {
             return MealType.valueOf(value.toUpperCase());
         } catch (IllegalArgumentException e) {
