+++ backend/src/main/java/com/fit/nufit/nutrient/dto/response/NutrientResponse.java
@@ -1,4 +1,4 @@
-package com.fit.nufit.nutrient.dto;
+package com.fit.nufit.nutrient.dto.response;
 
 import com.fit.nufit.nutrient.domain.Nutrient;
 import com.fit.nufit.nutrient.domain.NutrientUnit;
