+++ backend/src/test/java/com/fit/nufit/nutrient/domain/NutrientTest.java
@@ -3,8 +3,6 @@ package com.fit.nufit.nutrient.domain;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.test.annotation.Rollback;
 import org.springframework.transaction.annotation.Transactional;
 
 import static org.assertj.core.api.Assertions.*;
@@ -23,7 +21,7 @@ class NutrientTest {
 
         //when & then
         assertDoesNotThrow(()->{
-            new Nutrient("탄수화물", 4, NutrientUnit.of("g"));
+            new Nutrient("탄수화물", 4, NutrientUnit.from("g"));
         });
 
     }
